chore(reasoning): split modes.py into modes/ package — rule #5 debt cleanup
closure §8 option D: pre-existing rule #5 debt — core/reasoning/modes.py
was 30.0 KB, well over the CLAUDE.md 20 KB-per-core-file gate. Pure
structural split into the obvious 5-handler + 1-shared-constants shape
the file's own structure already advertised.

The split is purely structural: every handler body is byte-identical
to the pre-split version. The original module docstring's reasoning for
"free functions instead of methods" is preserved (now in modes/__init__.py).
All public import paths continue to work:

  # engine.py — UNCHANGED line-for-line
  from core.reasoning.modes import (
      handle_chat, handle_meta, handle_wiki_edit,
      handle_self_evolve, handle_coding,
  )

  # other tests: `import core.reasoning.modes as md` still gives you
  # md.handle_chat / md.CONTINUITY_DIRECTIVE_KO / etc.

Module sizes (CLAUDE.md rule #5 — 20 KB gate on core/):

  modes/__init__.py       2.3 KB  ✓  (re-export facade + docstring)
  modes/_common.py        1.7 KB  ✓  (CONTINUITY_DIRECTIVE_KO/EN)
  modes/chat.py           5.0 KB  ✓
  modes/coding.py         7.2 KB  ✓
  modes/meta.py           3.9 KB  ✓
  modes/self_evolve.py    8.1 KB  ✓
  modes/wiki_edit.py      5.9 KB  ✓
  Total                  34.1 KB     (was 30.0 KB monolith — +4 KB
                                      from package init / per-file
                                      docstrings + import overhead)

Largest file (self_evolve.py, 8.1 KB) is comfortably under 20 KB —
gives headroom for future tracing / log_stage wiring (Phase 0 L1
pattern) without pushing any submodule near the rule again.

Why a package instead of a flat splitting into modes_chat.py / etc.:

  - Existing call sites use `from core.reasoning.modes import handle_chat`
    style — a package with __init__.py preserves this verbatim.
  - `core.reasoning.modes` as a namespace stays self-describing
    (modes.chat / modes.coding etc. discoverable without a separate
    naming convention).
  - The package shape leaves room for future per-mode helper files
    (e.g. modes/wiki_edit/parsers.py if the action parser grows) without
    another top-level file proliferation.

Shared constants placement:

  CONTINUITY_DIRECTIVE_KO / CONTINUITY_DIRECTIVE_EN moved to
  modes/_common.py. Today only handle_chat consumes them; the original
  module-level comment notes that a future handler (handle_meta etc.)
  could re-use them without re-declaring. Keeping them in a shared file
  preserves that intent — and the leading underscore signals "package-
  private, not part of the public surface".

  __init__.py re-exports both constants so tests that probe
  ``core.reasoning.modes.CONTINUITY_DIRECTIVE_KO`` via hasattr continue
  to work without modification.

Test updates (2 files, surgical):

  tests/test_conversation_continuity.py
    ContinuityDirectiveTests.setUpClass previously set
    ``cls.src = inspect.getsource(_modes)``. After the split,
    ``inspect.getsource`` on the package returns only the __init__.py
    facade, which has no handle_chat body. Now points at
    ``inspect.getsource(modes.chat)`` for the two body-grep tests
    (test_handle_chat_gates_directive_on_current_session_history,
    test_directive_uses_korean_or_english_per_query). The hasattr +
    CONTINUITY_DIRECTIVE_* tests continue to work against
    ``cls.modes`` (the package) via the __init__.py re-export.

  tests/test_a2_phase2_selected_model.py
    ModesHandlerWiringTests._handler_body() previously sliced the
    monolith source by `def handle_X(` index. Replaced with
    ``inspect.getsource(getattr(md, name))`` which follows the
    function's __code__.co_filename — works identically whether the
    handler lives in monolith or submodule. Cleaner regardless of
    the split: doesn't depend on neighbouring functions' positions
    or `\n\ndef ` boundary heuristics.

No production source touched outside the modes split itself.

CLAUDE.md gate:
  #1 platform-only — pure refactor; no domain coupling
  #2 bench numbers — STEP 7 byte-identical (zero handler-body bytes
     changed; only the file each function lives in)
  #3 self-evolution — unrelated
  #4 architecture — no §5.* schema change; the §5 "modes dispatch"
     reference in ARCHITECTURE.md points at the package name, not the
     file, so still valid
  #5 module size — THIS IS THE FIX. All 7 files ≤ 8.1 KB ✓

Verification:
  - 91/91 modes-touching test suites green (test_conversation_continuity
    + test_response_style + test_coding_mode + test_meta_mode +
    test_external_role_isolation + test_a2_phase2_selected_model +
    test_reasoning_trace_helpers)
  - Full repo: 2041 passed; 3 pre-existing failures on main
    (test_character_prompt_cleanup x2 + test_character_remove_text_persona
    x1) — all character-feature tests, unrelated to modes split,
    confirmed by re-running with `git stash` reverting the split

Out of scope:
  - The 3 pre-existing character-test failures — separate follow-up
  - Future per-mode helper extraction (parsers, validators) — only
    worth doing when a mode grows beyond ~8 KB. Today none do.
